requirements.txt
argh==0.23.3
brewer2mpl==1.3.2
jedi==0.7.0
matplotlib==1.3.1
nose==1.3.0
numpy==1.8.0
protobuf==2.5.0
protobuf-to-dict==0.0.5
pyparsing==2.0.1
python-dateutil==2.2
scipy==0.13.1
six==1.4.1
tornado==3.1.1
unittest2==0.5.1
wsgiref==0.1.2
